Removed and renamed everything
