+++ src/main/java/it/unibz/pp2024/LudoKing/HelloController.java
@@ -1,4 +1,4 @@
-package it.unibz.wie2024.MarioParty;
+package it.unibz.pp2024.LudoKing;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
