+++ src/main/resources/it/unibz/pp2024/LudoKing/hello-view.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="it.unibz.wie2024.MarioParty.HelloController">
+      fx:controller="it.unibz.pp2024.LudoKing.HelloController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
